refactor: moved and renamed components
